fixes on explorer

+++ cogs/explorer.py
@@ -1,6 +1,7 @@
 import discord
 import random
 from discord.ext import commands
+from asyncio import TimeoutError
 
 
 class FileExplorer(commands.Cog, name="File Explorer (beta)"):
@@ -11,7 +12,7 @@ class FileExplorer(commands.Cog, name="File Explorer (beta)"):
     @commands.command()
     async def ls(self, ctx, path="~"):
         """List your files with the specified path"""
-        files = await self.db.find_one({"_id": ctx.author.id})
+        files = await self.get_dirs(ctx.author.id)
         if not path.startswith("/") and not path.startswith("~"):
             path = f"~/{path}"
 
@@ -24,7 +25,6 @@ class FileExplorer(commands.Cog, name="File Explorer (beta)"):
             )
             files = {"_id": ctx.author.id, "home": {}}
 
-        files.pop("_id")
         tg = [x for x in path.split("/") if x]
         if path.startswith("/") and len(tg) != 0:
             tg[0] = f"/{tg[0]}"
@@ -58,8 +58,7 @@ class FileExplorer(commands.Cog, name="File Explorer (beta)"):
     @commands.command()
     async def cat(self, ctx, file):
         """Get the content of a file"""
-        files = await self.db.find_one({"_id": ctx.author.id})
-        files.pop("_id")
+        files = await self.get_dirs(ctx.author.id)
         if not file.startswith("/"):
             file = f"~/{file}"
 
@@ -75,6 +74,10 @@ class FileExplorer(commands.Cog, name="File Explorer (beta)"):
 
         if isinstance(files, bytes):
             files = files.decode("UTF-8")
+        
+        if files == "":
+            return await ctx.send("File is empty.")
+
         await ctx.send(
             f"```\n{files}\n```" if not isinstance(files, dict) else "Is a directory"
         )
@@ -82,8 +85,7 @@ class FileExplorer(commands.Cog, name="File Explorer (beta)"):
     @commands.command(usage="<file1> [file2] [file3] ...")
     async def touch(self, ctx, *files):
         """Create a file"""
-        _files = await self.db.find_one({"_id": ctx.author.id})
-        _files.pop("_id")
+        _files = await self.get_dirs(ctx.author.id)
         paths = []
         if not files:
             raise commands.MissingRequiredArgument(
@@ -126,8 +128,7 @@ class FileExplorer(commands.Cog, name="File Explorer (beta)"):
     async def rm(self, ctx, *files):
         """Delete a file"""
         files = list(files)
-        _files = await self.db.find_one({"_id": ctx.author.id})
-        _files.pop("_id")
+        _files = await self.get_dirs(ctx.author.id)
         _dirs = "-r" in files or "-rf" in files or "-fr" in files
         _force = "-f" in files or "-rf" in files or "-fr" in files
         _no_preserve_root = "--no-preserve-root" in files
@@ -202,25 +203,24 @@ class FileExplorer(commands.Cog, name="File Explorer (beta)"):
                         title="Wait, that's dangerous!",
                         description=(
                             "This will delete **ALL**  your data you saved here.\n"
-                            'Please send "Yes, I am aware that this is a very bad idea and I will lose all my data.", if you are.'
+                            'Please send "Yes, I know this is a very bad idea and I am going to lose all my data.", if you are.'
                         ),
                         color=discord.Color.red(),
                     )
                 )
 
                 def msgchk(m):
-                    return (
-                        m.author.id == ctx.author.id
-                        and m.content
-                        == "Yes, I am aware that this is a very bad idea and I will lose all my data."
-                    )
+                    return m.author.id == ctx.author.id
 
                 try:
                     msg = await self.bot.wait_for("message", check=msgchk, timeout=20)
                 except TimeoutError:
-                    await message.edit(content="Aborted.")
+                    return await message.edit(content="Aborted.")
                 else:
-                    await self.db.delete_one({"_id": ctx.author.id})
+                    if msg.content == "Yes, I know this is a very bad idea and I am going to lose all my data.":
+                        await self.db.delete_one({"_id": ctx.author.id})
+                    else:
+                        return await message.edit(content="Aborted.")
             else:
                 del files_[tg[-1]]
                 await self.db.find_one_and_update(
@@ -233,6 +233,61 @@ class FileExplorer(commands.Cog, name="File Explorer (beta)"):
     async def edit(self, ctx, file):
         await ctx.send("Coming late®!")
 
+    @commands.command(usage="<dir1> [dir2] [dir3] ...")
+    async def mkdir(self, ctx, *files):
+        """Create a file"""
+        _files = await self.get_dirs(ctx.author.id)
+        paths = []
+        if not files:
+            raise commands.MissingRequiredArgument(
+                type("testù" + ("ù" * 100), (object,), {"name": "dir1"})()
+            )
+        for _file in files:
+            file = _file
+            files_ = _files
+            if not file.startswith("/"):
+                file = f"~/{file}"
+
+            file = file.replace("~", "/home")
+            tg = [x for x in file.split("/") if x]
+
+            for dir in tg:
+                if dir == tg[-1]:
+                    break
+                try:
+                    files_ = files_[dir.replace("/", "", 1)]
+                except KeyError:
+                    return await ctx.send("No such file or directory")
+
+            if tg[-1] in files_:
+                if isinstance(files_[tg[-1]], dict):
+                    return await ctx.send(f"Error creating `{file}`: Directory exists.")
+                return await ctx.send(f"Error creating `{file}`: File exists.")
+
+            files_[tg[-1]] = {}
+            paths.append(file + "/")
+            await self.db.find_one_and_update({"_id": ctx.author.id}, {"$set": _files})
+
+        if len(paths) == 1:
+            msg = f"Done. You can find your directory at `{paths[0]}`"
+        else:
+            ns = "\n".join(paths)
+            msg = f"Done. You can find your directories at: **```fix\n{ns}\n```**"
+        await ctx.send(msg)
+
+    async def get_dirs(self, id):
+        ret = await self.db.find_one({"_id": id})
+        if ret == None:
+            await self.db.find_one_and_update(
+                {"_id": id},
+                {"$set": {"home": {}}},
+                upsert=True,
+            )
+            ret = await self.db.find_one({"_id": id})
+        
+        ret.pop("_id")
+
+        return ret
 
 def setup(bot):
     bot.add_cog(FileExplorer(bot))
